requirements.txt
flask
python-dotenv
langchain
openai
faiss-cpu
pytesseract
pillow
pypdf2
python-docx
pandas
plotly
moviepy
whisper
